Added a functionality for the user to specify if they want the tool to count the amount of genes that were counted for every sample. The output of this count is written to a text file in the output archive.
Added a functionality for the user to specify if they want the tool to count the amount of genes that were counted for every sample. The output of this count is written to a text file in the output archive. An additional count per locus could theoretically be added in the future appending to this new file.