Replaced very specific lines of code in the python script that caused overfitting
Prevented overfitting on test dataset
Specifically in the python script:

Replaced a line of code that checked for files with the name "gene_*" to instead just check if the file is a directory or not.

Replaced the line that removes the word 'gene' to instead remove every letter from the string.